Add fix for #2702

+++ Setup/update.sh
@@ -63,9 +63,18 @@ fi
 echo "***************     Ericom Shield Update ($CONTAINER_TAG, $ARGS $ES_CHANNEL) ..."
 
 echo "$(date): Ericom Shield Update: Running Update" >>"$LOGFILE"
-docker run --rm -it \
-   -v /var/run/docker.sock:/var/run/docker.sock \
-   -v $(which docker):/usr/bin/docker \
-   -v /usr/local/ericomshield:/usr/local/ericomshield \
-   -e "ES_PRE_CHECK_FILE=$ES_PRE_CHECK_FILE" \
-    "securebrowsing/$CONTAINER_TAG" $ARGS $ES_CHANNEL
+if ! [[ $0 != "$BASH_SOURCE" ]]; then
+    docker run --rm  -it \
+       -v /var/run/docker.sock:/var/run/docker.sock \
+       -v $(which docker):/usr/bin/docker \
+       -v /usr/local/ericomshield:/usr/local/ericomshield \
+       -e "ES_PRE_CHECK_FILE=$ES_PRE_CHECK_FILE" \
+        "securebrowsing/$CONTAINER_TAG" $ARGS $ES_CHANNEL
+else
+    docker run --rm \
+       -v /var/run/docker.sock:/var/run/docker.sock \
+       -v $(which docker):/usr/bin/docker \
+       -v /usr/local/ericomshield:/usr/local/ericomshield \
+       -e "ES_PRE_CHECK_FILE=$ES_PRE_CHECK_FILE" \
+        "securebrowsing/$CONTAINER_TAG" $ARGS $ES_CHANNEL
+fi
